refactor(scripts): modularize functionality
